fix bug where unread equals total
This bug occurs when the request to get the archived count fails, but
the call to get total count succeeds. Then our calculated unread count
is total minus zero, because archived is initialized to zero. In this
case our unread count is equal to total count. This is bug and a big
peak in the graph.

The fix checks for unread being equal to total and archvied being zero
and, most important, last archived count (from the call before) being
greater than zero.

It is very unlikely that this case happens in a real scenario, because
the user would have to mark all articles as not read although he read
them in past. So this would be the reverse process of the main use case
of wallabag, something like reverse reading or unreading.